feat(40) canary deployment done

+++ sdk/java/sdk/src/main/java/ru/itmo/config_streamer/sdk/Config.java
@@ -1,6 +1,6 @@
 package ru.itmo.config_streamer.sdk;
 
-public record Config(String key, int version, Object payload) implements Cloneable {
+public record Config(String key, long version, Object payload) implements Cloneable {
     @Override
     public Config clone() {
         return new Config(key, version, payload);
